Points of Interest, needs more work on formatting and stuff
Add Point of Interest field to MMBRoom

While drawing the image, number the points of interest, label them beneath the room name as the number
Make the image bigger by the height required to display the points of interest text at the bottom.
Draw the points of interest text at the bottom.